OpenCL histograms and means working. Slight difference with CPU version. OpenCL helper functions added to Tobago. Slight bug found with frame numbers and time stamps that affects all code, needs to be fixed.

+++ Synesthesia/Tobago/utility/Colors.h
@@ -5,15 +5,15 @@
 #include <iostream>
 
 typedef struct {
-    float r;       // percent
-    float g;       // percent
-    float b;       // percent
+    double r;       // percent
+    double g;       // percent
+    double b;       // percent
 } rgb;
 
 typedef struct {
-    float h;       // angle in degrees
-    float s;       // percent
-    float v;       // percent
+    double h;       // angle in degrees
+    double s;       // percent
+    double v;       // percent
 } hsv;
 
 typedef struct {
